Improve unit tests, and coverage, for Control's registerLogHandler
